+++ substrafl/algorithms/pytorch/torch_fed_avg_algo.py
@@ -63,8 +63,8 @@ class TorchFedAvgAlgo(TorchAlgo):
                 ):
 
                     # Create torch dataloader from the automatically instantiated dataset
-                    # ``train_dataset = self._dataset(datasamples=datasamples, is_inference=False)`` is executed prior
-                    #  the execution of this function
+                    # ``train_dataset = self._dataset(data_from_opener=data_from_opener, is_inference=False)``
+                    # is executed prior the execution of this function
                     train_data_loader = torch.utils.data.DataLoader(train_dataset, batch_sampler=self._index_generator)
 
                     for x_batch, y_batch in train_data_loader:
@@ -153,7 +153,7 @@ class TorchFedAvgAlgo(TorchAlgo):
     @remote_data
     def train(
         self,
-        datasamples: Any,
+        data_from_opener: Any,
         shared_state: Optional[FedAvgAveragedState] = None,  # Set to None per default for clarity reason as
         # the decorator will do it if the arg shared_state is not passed.
     ) -> FedAvgSharedState:
@@ -166,7 +166,7 @@ class TorchFedAvgAlgo(TorchAlgo):
             * compute the weight update
 
         Args:
-            datasamples (typing.Any): Input data returned by the ``get_data`` method from the opener.
+            data_from_opener (typing.Any): Input data returned by the ``get_data`` method from the opener.
             shared_state (FedAvgAveragedState, Optional): Dict containing torch parameters that
                 will be set to the model. Defaults to None.
 
@@ -176,7 +176,7 @@ class TorchFedAvgAlgo(TorchAlgo):
         """
 
         # Create torch dataset
-        train_dataset = self._dataset(datasamples, is_inference=False)
+        train_dataset = self._dataset(data_from_opener, is_inference=False)
         if shared_state is None:
             # Instantiate the index_generator
             assert self._index_generator.n_samples is None