+++ substrafl/algorithms/pytorch/torch_newton_raphson_algo.py
@@ -279,7 +279,7 @@ class TorchNewtonRaphsonAlgo(TorchAlgo):
     @remote_data
     def train(
         self,
-        datasamples: Any,
+        data_from_opener: Any,
         # Set shared_state to None per default for clarity reason as
         # the decorator will do it if the arg shared_state is not passed.
         shared_state: Optional[NewtonRaphsonAveragedStates] = None,
@@ -296,7 +296,7 @@ class TorchNewtonRaphsonAlgo(TorchAlgo):
             * a L2 regularization can be applied to the loss by settings `l2_coeff` different to zero (default value)
 
         Args:
-            datasamples (typing.Any): Input data returned by the ``get_data`` method from the opener.
+            data_from_opener (typing.Any): Input data returned by the ``get_data`` method from the opener.
             shared_state (NewtonRaphsonAveragedStates, Optional): Dict containing torch parameters that
                 will be set to the model. Defaults to None.
 
@@ -308,7 +308,7 @@ class TorchNewtonRaphsonAlgo(TorchAlgo):
         """
 
         # Create torch dataset
-        train_dataset = self._dataset(datasamples, is_inference=False)
+        train_dataset = self._dataset(data_from_opener, is_inference=False)
 
         if shared_state is None:
             # Instantiate the index_generator